Disables Zig Zag printing
Splatoon 3 seems to have some small issue with the Zig Zag printing (making the cursor go left to right on one line, then right to left on the next). It was made to save time back when but it's not working properly yet. I left a commented line with the command with the option, if you want to try.